Added a method for removing an item from the cart
The `RemoveItemFromCart` method has been implemented in CartsController.cs, which allows you to remove items from the user's cart. The method checks for the user's presence, loads the cart, searches for and removes the specified item, updates the last modification time, and saves the changes to the database. It returns an updated view of the cart in DTO format.