Improved validation and updating of DTO in the cart
Changes have been made to CartsController.cs to add checks for model validity and animal existence when creating an item in the cart. The update method now accepts a CartItemUpdateDTO object, which simplifies data transfer. Checks have been added for the existence of the user, cart, and item in the cart.

The way the list of items is created in CartDTO.cs has been changed to use the new CartItemDTO format. Validation attributes have been added to the Quantity and AnimalId properties in CartItemDTO.cs. A new CartItemUpdateDTO class has been created with similar validation attributes.