Added order controller and improved cart management
Changes have been made to `CartsController.cs` to suppress warnings about possible `null` returns.
Added a new `OrdersController` controller with routing and authorization.
Methods have been implemented to retrieve all orders and orders by user ID, as well as helper methods to check the existence of an order and retrieve a valid user ID.
Entity Framework is used to work with the database and the Bearer scheme is used for authorization.