Added CreatedAt and UpdatedAt properties to the model
Changes have been made to the `TameShopDbContext` database context, adding `CreatedAt` and `UpdatedAt` properties with default values for the `Cart`, `CartItem`, and `Order` entities. The corresponding properties in `UserDbContextModelSnapshot.cs` have been updated. Removed initializations of `DateTime.UtcNow` values in the `Cart` and `CartItem` models. Added dependencies for repositories and services in `Program.cs`. Changed the logic in `CartsService` to match the new models. Created the `datescreation` migration to update the columns in the database.